fix:
arrumando problema de sintax de sql

+++ src/main/resources/db/migration/V5__Create_Table_Pessoa_Endereco.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS `pessoa_endereco` (
-    `pessoa_id` bigint(20) NOT NULL,
-    `endereco_id` bigint(20) NOT NULL,
+    `pessoa_id` INTEGER NOT NULL,
+    `endereco_id` INTEGER NOT NULL,
     PRIMARY KEY(pessoa_id, endereco_id) 
 )
